experiments/intelligent-video-search-ai-agent/requirements.txt
mcp==1.10.1
opensearch-py==2.4.2
requests-aws4auth==1.2.3
numpy==1.24.3
python-dotenv==1.0.0
boto3>=1.34.0
strands-agents
strands-agents-tools
fastapi
uvicorn[standard]
websockets>=11.0.3
requests
python-multipart
httpx>=0.25.2
anyio>=3.7.1
twelvelabs>=0.2.0




